+++ examples/get_fun.R
@@ -1,0 +1,7 @@
+get_fun(plu_ral)
+get_fun(plu::ral)
+get_fun("plu_ral")
+get_fun("plu::ral")
+
+get_fun(NULL)
+get_fun(NULL, default = plu_ral)